backend/requirements.txt
Django
djangorestframework
channels
django-cors-headers
asgiref
daphne
dotenv
psycopg
django-jazzmin